Specify grammar in textX
I decided to let the parser just look for a list of integers for
rule triggers as I don't think I can have the parser look for a
subtype (naturals) while still spitting a number out to python.
I could change the definition of trigger from INT to /[1-9][0-9]*/
but then the parser will spit it out as a string instead of
being nice and spitting out a number. I'd like to see how this ends
up looking in the context of the rest of the interpreter. I may end
up changing to the above regex instead.

textX homepage: http://www.igordejanovic.net/textX/